Fix Swagger schema for FormRedirect
The reason it was broken is that genericDeclareNamedSchema tries to define the schema for this type as a heterogeneous array (i.e. tuple) with Swagger types String and AuthnRequest. However, Swagger does not support heterogeneous arrays, and this results in an array whose underlying type which is at the same time marked as a string, and referring to the schema for AuthnRequest, which is of course invalid.