Expand out LegalholdProtectee'
The `LegalholdProtectee'` type does not actually have anything to do with legal hold. It is simply a pair of a `UserId` and a "user type", i.e. a bit representing whether the user id refers to an actual user or a bot.

This removes the LegalholdProtectee' type and replaces it with the explicit pair.