Refactor qualified send message endpoint (#1634)
* Finish the function to check message clients

* Create event from qualified message

* Refactor Message module

* Introduce UnionT

* Restore legalhold policy conflict check

* Generate user and bot notifications

* Complete definition of postQualifiedOtrMessage

* Hook refactored message sending to API

* Expand out LegalholdProtectee'

The `LegalholdProtectee'` type does not actually have anything to do with legal hold. It is simply a pair of a `UserId` and a "user type", i.e. a bit representing whether the user id refers to an actual user or a bot.

This removes the LegalholdProtectee' type and replaces it with the explicit pair.

* Refactoring: move definitions out of Update

* Ignore sender when checking validity of clients

* Work with flattened sets of triples throughout

* Do not report ignored missing clients

* Simplify sender client exclusion logic

* Disable redundant check in integration tests

With this, integration tests should pass

* Switch to the new implementation of msg sending

Hooked the old unqualified endpoint for message sending to the new
implementation.

Also, the checkMessageClients function now takes a map of client sets
instead of a set of triples. This is to distinguish the case where a
user is present in a conversation but has no clients.

* Restore sender client checks

* Move toBase64Text to Data.Json.Util

* Update CHANGELOG

* Unqualify response using htrans

This replaces the manual pattern matching in terms of explicit natural
numbers with a type-class based approach, and uses the htrans function
with an explicit constraint to perform the conversion inside the NS
structure.

Co-authored-by: Akshay Mankar <akshay@wire.com>