Remove unused HEAD endpoints
Servant implements HEAD endpoints automatically in terms of GET
endpoints. As it turned out, the explicit implementations we had for
`HEAD /users/:uid` (and its qualified counterpart) ended up not being
invoked by Servant's routing mechanism.

This simply removes the endpoints, and adds tests showing that the
functionality is still present and works as expected.