Use a different certificate for federator
This makes it less awkward to use the server certificate for the
federator ingress also as a client certificate for requests to other
federators. Otherwise, the client certificate would contain extraneous
domains.

This also uses Let's Encrypt intemediate CA (from their staging
environment) as a default value for the CA in the secret.