Add a mechanism for deriving `AsUnion` instances
This is based on an ad-hoc type class `AsConstructor xs r`, where `r` is
a response and `xs` is a list of types such that `r` is convertible to
the product of the `xs`.

The `AsConstructor` class can be instantiated to get automatic
derivation of `AsUnion` instances via `GenericAsUnion`. The idea is that
one has to make sure that for each response `r` in a `MultiVerb`
endpoint, there is an instance of `AsConstructor xs r` for some `xs`,
and that the list `xss` of all the corresponding `xs` is equal to
`GSOP.Code` of the handler type. Then one can write:

```haskell
  type Responses = ...
  data Result = ...
    deriving stock (Generic)
    deriving (AsUnion Responses) via (GenericAsUnion Responses Result)

  instance GSOP.Generic Result
```

and get an `AsUnion` instance for free.

There are a few predefined instances for constructors taking a single
type corresponding to a simple response, and for empty responses, but in
more general cases one either has to define an `AsConstructor` instance
by hand, or derive it via `GenericAsConstructor`.

This commit also defines convenience functions `eitherToUnion` and
`eitherFromUnion` that can be used to define an instance of `AsUnion`
for a type of the form `Either a b`, given instances for `a` and `b`.