Separate ca secret from tls secret
It seems cert-manager takes control of the `ca.crt` key in the secret
corresponding to a certificate, so it does not seem possible to set it
manually when cert-manager is enabled. Unfortunately, it still needs to
be a secret, since the ingress requires it.

This commit moves the CA to a separate secret. Therefore, it needs to be
set explicitly in the chart values for both federator and the
nginx-ingress-services, even if federator secret sharing is enabled.