Finish integrating `MultiVerb` (#1693)
This PR completely removes `UVerb` from the codebase, as well as the `EmptyResult` kludge that we needed to generate empty responses. Both are now replaced by `MultiVerb`.

It also adds a mechanism for deriving `AsUnion` instances, as well as convenience functions  `eitherToUnion` and
`eitherFromUnion` that can be used to define an instance of `AsUnion` for a type of the form `Either a b`.