Ensure clients only recieve messages meant for them in remote convs
This commit removes duplicated logic to push messages to clients by making logic
in Galley.API.Message.sendLocalMessages accomodating to pushing messages for
remote conversations, but still enforcing the recipients to be local. Using this
function from federation endpoint of galley allows us to not duplicate the logic
for how a message becomes an event and gets pushed to each client.

The commit also moves this logic from Galley.API.Update to Galley.API.Federation
as most of the logic to translate from inputs of the federation endpoint to
inputs for Galley.API.Message.sendLocalMessages is very federation specific and
doesn't need to live in the public API module.