Revert hack to display helm test output
The hack in #1700 does not actually seem to work in CI, and in fact it
prevents log output from being displayed at all when helm test fails.

The integration-test-logs.sh script is still around, and it can be used
to display helm test output interactive when running tests manually in a
terminal.