Remove qualified GET conversation self
This was a newly added qualified endpoint stub, which raised an error
when used on a remote conversation.

We are now deprecating fetching conversation self membership information
altogether (since clients do not use it, and they can always get that
information from the conversation itself), so this endpoint can simply
be removed.