Extract `MemberStatus` from `LocalMember`
This is a refactoring of `LocalMember` to extract the member status
fields (hidden/muted/archived) to their own type, so that they can later
be reused to transform a remote conversation to a local one.

Also, the `InternalMember` type constructor has been removed, since the
only used instantiation of it was the `LocalMember` type synomyn.