Fix remote user self info in get-conversations RPC
This removes the hack of returning a fake self member in the
get-conversations RPC, by introducing a different type for remote
conversations.

A `RemoteConversation` object contains the same
information as a normal `Conversation`, except the self member is
replaced by simply a `RoleName`, which is the only bit of information
for remote users that is actually stored locally.

The functions in `Galley.API.Mapping` have been rewritten to fit the new
convention.

Notes:

 - the Mapping tests have been commented out; they probably should be
   rewritten and become unit tests;

 - remote conversation status flags are still always set to their
   default values.