Remove otr_muted from Member and related types (#1751)
* Remove otr_muted from Member and related types

This field has been deprecated for some time, and the clients do not
need it. Note that the corresponding conversation field in the database
is still present, and removing that is left to future work.

* Restore muted status assertions

Before these were assertions on the `otr_muted` field, but now that
field has been removed, so we should assert on `otr_muted_status`
instead.