Test hotfix.
There is an error in this test if we do it the way it's supposed to
work, but if we move an offending entry out of the way in
`spar.issure_idp_v2` before running into the bug, the test passes.

Revert of this commit and proper fix coming up!