Make sure that unnamed schemas are referenced (#1802)
This changes the Swagger implementation of `unnamed` in
schema-profunctor to drop the name, but still declare the corresponding
named schema. Now even unnamed schemas that are not actually referenced
anywhere will appear in the list of schemas in Swagger.

Also changed the name of the schema for `SimpleMember` to make it
reflect the Haskell type name.

Thanks @atomrc for reporting this issue.