Fix check for leave action
A leave action is a special case of a member remove action, but it has a
different "tag", so now `conversationActionTag` takes the originating
user as an extra argument, and retuns a `LeaveConversation` tag for
`ConversationActionRemoveMember` actions that remove the originating
user.

This also fixes legalhold-related failures when non-consenting users are
removed after legal hold is enabled.