Honor `defaultIfNew`.
I forgot to actually write this default into newly created users if
available.

Detail: if you want `defaultIfNew` to be `null` (so you can change
your mind about teh default for everybody later), just omit the line
in the yaml file.  This wasn't possible before, as there is also a
hard-coded default in the source.