Introduce fine-grained placeholder effects
This commit introduces several placeholder effects, mostly having to do
with making HTTP requests. All the existing uses of `MonadUnliftIO` are
now either gone, or hidden behind of of these effects, and that made it
possible to get rid of the `MonadUnliftIO` instance for `Galley`.

Also, the `Galley0` type synomym now refers to `Galley` without any
effects, so `runGalley` and related functions now take a `Galley
GalleyEffects` instead.

`Galley0` still has a `MonadUnliftIO` instance, so it can be used as a
temporary crutch to get access to async primitives. Those need to be run
in `Galley0`, and finally lifted to a general `Galley r` monad.
Eventually, the `Galley0` actions will simply be replaced by effect
actions, and the code actually using `MonadUnliftIO` will be relegated
to interpreters.