Optimise remote user deletion (#1872)
Creates two Federation RPCs:

* In brig: on-user-deleted, notify about the connections in chunks of 1000 users.
* In galley: on-user-deleted, notify about the conversations in chunks 1000 conversations

When writing integration tests in brig, we can mock the federator for brig but not galley. As the two RPCs must be made from two separate places. So, we had to mock out galley to be able to test the brig functionality. The galley functionality is tested separately by calling the internal endpoint.

Co-authored-by: Akshay Mankar <akshay@wire.com>
Co-authored-by: Stefan Matting <stefan@wire.com>