Galley polysemy (1/5) - Introduce Sem and "access" effects (#1881)
* Add type variable to Galley monad

This is step 0 in the process of converting galley to effects. We
introduce a phantom type variable `r` in the `Galley` monad, which will
later be used for the effect row.

* Use API instead of DB access in 1-1 conv test

* Monomorphise Data functions

* Avoid MonadUnliftIO in Bilge.RPC

* Remove unneeded MonadLogger constraint

* Introduce fine-grained placeholder effects

This commit introduces several placeholder effects, mostly having to do
with making HTTP requests. All the existing uses of `MonadUnliftIO` are
now either gone, or hidden behind of of these effects, and that made it
possible to get rid of the `MonadUnliftIO` instance for `Galley`.

Also, the `Galley0` type synomym now refers to `Galley` without any
effects, so `runGalley` and related functions now take a `Galley
GalleyEffects` instead.

`Galley0` still has a `MonadUnliftIO` instance, so it can be used as a
temporary crutch to get access to async primitives. Those need to be run
in `Galley0`, and finally lifted to a general `Galley r` monad.
Eventually, the `Galley0` actions will simply be replaced by effect
actions, and the code actually using `MonadUnliftIO` will be relegated
to interpreters.

* Remove MonadMask instance of Galley

This also introduces a `SparAccess` effect and adds a few more
`BrigAccess` and `BotAccess` constraints.

* Remove MonadCatch instance of Galley

* Turn Galley into a Sem newtype

The underlying `Sem` monad in `Galley` is an arbitrary effect stack that
contains at least the effects which replicate the functionality of the
original `Galley` monad. All the functionality has been reimplemented in
terms of `Sem`, so the existing code does not need to be changed at all.