Galley polysemy (2/5) - Store effects (#1890)
* Introduce ConversationStore effect

An effect for creating and deleting conversations.

This commit also moves the corresponding Cassandra-specific code to the
`Galley.Cassandra` hierarchy. This is for better module organisation:
effects should depend on `Data` modules (for action types), but not on
the specific implementation of these effects. It also avoids include
cycles.

* Add Paging module and ListItems effect

* Add MemberStore effect

* Add TeamStore effect

* Add ListItems effect for teams

* Add TeamMemberStore effect

* Add ClientStore effect

* Add CodeStore effect

* Remove Data module

* Add CHANGELOG entry

* Add ServiceStore effect

Also add `createBotMember` action to `MemberStore`.

Co-authored-by: Marko Dimjašević <marko.dimjasevic@wire.com>