charts/wire-server-metrics: Use kube-prometheus-stack (#1915)
The old version of prometheus operator chart is no longer compatible with latest
K8s versions. The chart has also since moved to kube-prometheus-stack on the
prometheus-community helm chart reposistory from the helm-stable repository.

The kube-prometheus-stack helm chart also allows creating dashboard using labels
on configmaps, so now we don't have to refer to non-pinned versions of the
dashboard definitions. Instead each dashboard now becomes a configmap with a
specific label and grafana discovers it.

This commit also updates the dashboards to use new ways of addressing the pods
which came with the move from old version of prometheus-operator chart to the
kube-prometheus-stack chart. Along with these changes, the dashboards now also
use the qualified endpoints and federation endpoints to display information
about prekey claims and message sending stats.