Change federation API to be based on HTTP2
This is a preliminary commit that removes gRPC completely from
wire-api-federation, and assumes that federator is going to be accepting and
forwarding HTTP2 requests directly. Failures are going to be propagated using
normal HTTP error codes, and using Wai.Error values as responses.

The federation API is still following the same conventions as before, with the
exception that now we require that every endpoint consists of a single path
segment, which is the name of the RPC. For this reason, the `on-user-deleted`
endpoints in brig and galley have been updated to single-segment endpoints.

 - Removed gRPC dependency.
 - Removed explicit "/federation" prefix from federation APIs.
 - `clientRoutes` is now a polymorphic value that works for both services.
 - Brig and Galley API modules can now be imported without clashes.
 - Federator client has been completely rewritten, and now implements HTTP2
 client functionality on top of the http2 library. The low-level http2 client
 functionality that establishes a connection supports both plaintext and TLS,
 and will also be used to implement the outward service of federator.
 - Federator client errors have been restructured and documented. They are
 organised into three layers: low-level connection/TLS errors, federator client
 errors, and high-level errors thrown by the application code.
 - The origin domain header type in Servant is now a special combinator that
 works just like a header on the server side, but is removed from the resulting
 client type. Federator client knows to add the origin domain header on every
 request, whether or not the origin domain is present in the API type.