Add new LogFormat: StructuredJSON
This format is intended as an imporvement over the JSON format. It tries to
improve in these ways:

- Explicitly encode log level as a key in the main JSON object logged

- Encode field values as key-value pairs in the main JSON object

Encoding logs like this makes it easy for external tools like fluent-bit and
elasticsearch to understand the log level and the fields and index them as such
without requiring bonanza to be present while processing logs.