Replace gRPC with HTTP2 in federator (#1930)
* Change federation API to be based on HTTP2

This is a preliminary commit that removes gRPC completely from
wire-api-federation, and assumes that federator is going to be accepting and
forwarding HTTP2 requests directly. Failures are going to be propagated using
normal HTTP error codes, and using Wai.Error values as responses.

The federation API is still following the same conventions as before, with the
exception that now we require that every endpoint consists of a single path
segment, which is the name of the RPC. For this reason, the `on-user-deleted`
endpoints in brig and galley have been updated to single-segment endpoints.

 - Removed gRPC dependency.
 - Removed explicit "/federation" prefix from federation APIs.
 - `clientRoutes` is now a polymorphic value that works for both services.
 - Brig and Galley API modules can now be imported without clashes.
 - Federator client has been completely rewritten, and now implements HTTP2
 client functionality on top of the http2 library. The low-level http2 client
 functionality that establishes a connection supports both plaintext and TLS,
 and will also be used to implement the outward service of federator.
 - Federator client errors have been restructured and documented. They are
 organised into three layers: low-level connection/TLS errors, federator client
 errors, and high-level errors thrown by the application code.
 - The origin domain header type in Servant is now a special combinator that
 works just like a header on the server side, but is removed from the resulting
 client type. Federator client knows to add the origin domain header on every
 request, whether or not the origin domain is present in the API type.

* Rewrite federator services as HTTP applications

This is a re-implementation of the inward and outward federator services as
simple Wai applications, with the following interface:

 * federation client / outward service path
      /rpc/:domain/:component/:rpc

 * inward ingress path
      /federation/:component/:rpc

The origin domain is passed as a header directly in the form expected by the
federation API in brig and galley (i.e. a `Wire-Origin-Domain` header).
Server-to-server authentication is not affected, and the client certificate is
received by the inward service as a header, exacly as it was before.

 - Removed gRPC dependency.
 - Both the inward and outward services are now simple Wai applications that
 serve a single route each, and forward the request appropriately, without
 looking at the body.
 - Removed error layer from the main application monad.
 - Renamed `LookupError` to `DiscoveryFailure` for consistency.
 - All polysemy error effects are now converted to Wai responses with a JSON
 `Wai.Error` as body.
 - Restored and simplified the federator mock server. Its functionality is
 similar to before the change, but the gRPC type for federated requests has
 been replaced by a simpler custom type `FederatedRequest`, which is used only
 by the mock machinery itself.
 - The internal Polysemy effects used in federator have been re-organised
 slightly. For example, the `Remote` effect never fails by itself, and
 `RemoteError` can be thrown by its interpretation, instead. Also, error
 handling and logging is performed in a single place (`Federator.Response`) for
 both services.
 - The canonical `Remote` effect interpretation is now using the low-level
 HTTP2 client functionality of Federator client to forward requests to a remote
 federator.
 - Validation has been simplified. In particular, path sanitisation has been
 removed, since the RPC path is now a normal HTTP path, which is assumed to
 have already been validated and sanitised by the HTTP server (and probably all
 the intermediate proxies and load balancers).
 - The federator client tests that were residing in wire-api-federation before
 have now been moved to the federator package, since they depend on the mock
 federator code.

* Brig: follow changes in federator

Adapted brig to the changes in federator and the corresponding mock
machinery. This is mostly about following type changes, and installing
the federation API at its correct "/federation" prefix, since the prefix
has now been removed from the Servant API types.

* Galley: follow changes in federator

Adapted galley to the changes in federator and the corresponding mock
machinery. This is mostly about following type changes, and installing
the federation API at its correct "/federation" prefix, since the prefix
has now been removed from the Servant API types.

* Replace gRPC with HTTP in ingress configuration

* Makefile improvements

* Add CHANGELOG entry

* Follow http2 upstream preface-race branch

* HTTP2 client: test parsing of response header

* Remove all remaining mentions of gRPC

 - Delete obsolete python federation client
 - Delete out-of-date federation documentation
 - Change mentions of gRPC to HTTP2 where appropriate

* Restore federation-not-available label

* Turn all immediate federator errors into 403

* Map remote federator error responses

When the remote federator returns an error response, fail by returning a
533 error which is specific for this situation. This makes it possible
to distinguish local and remote federation failures, and makes sure that
a 5xx error is always returned for remote ones.

* Convert local 403 federator errors to 500

If the local federator returns 403, that indicates an issue in federator
itself, or in the communication between the service making a federated
call and the local federator. This should result in a 500 status code.

* remove commented-out code

* ExternalServer enforces RPC name format

Co-authored-by: Paolo Capriotti <paolo@capriotti.io>