Add shell.nix for ad-hoc development environments
There seems to be no direnv plugin for VSCode that works with remotes.
(I haven't checked the local / non-remote case.)
This commit introduces a shell.nix file that depends on the same package
list as direnv.nix, but uses mkShell instead of buildEnv to create the
environment. This works fine with nix-shell itself and VSCode's
'Nix Environment Selector' plugin.