Update federation error documentation (#1956)
* Document general federation error mapping strategy

* Move swagger header to a separate file

The markdown file is converted to a haskell module by a custom
`Setup.hs`.

* Return 400 error code for federation-denied

The federation-denied error is client-visible, since it may be due to
errors in the original request. This makes the error propagate to the
client with status code 400.

* Update federation error documentation

Also change some error labels and status codes to make them more consistent.

The authentication error category has been merged with the remote
federation error category, and all errors of the same category now have
similar status codes.