charts/nginz: fix fail-open behaviour in CORS configuration.
The previous commits adding CORS configuration set the access control header
to unconditionally return the Origin header sent with the request. This would
cause fail-open behaviour, where any Origin sent by a client would be allowed.

Instead, the $cors_header variable is used, as this is specifically set based
on the request Origin header so that only origins which are explicitly in the
Helm chart's allow list configuration may make cross-origin requests to nginz
API endpoints.