Replace servant-generic with Named combinator (#2022)
The generic mechanism in Servant for defining API types is cumbersome
and slow to compile. This commit replaces the API record for the public
Galley API with a tree explicitly built using the `:<|>` operator.

To make up for the lack of record labels, it also introduces a `Named`
combinator for servant API types, which can be used to tag an endpoint
with a name (usually a symbol). The name must match the one with which
the handler is wrapped.