Brig: Servantify self API (#2091)
* Brig: Servantify `PUT /self` endpoint

* Brig: Servantify PUT `/self/phone` endpoint

* Brig: Sevantify `DELETE /self/{email,phone}` endpoints

* Brig: Servantify `HEAD /self/password` endpoint

Also add instnce for AsUnion '[..] Bool back.

* Brig: Servantify `PUT /self/password` endpoint

* PasswordChange: Adjust golden JSONs to not include null old_password

Only `FromJSON` needs to be backwards compatible as this object is consumed as
request body by brig. So, breaking `ToJSON` like this shouldn't cause any issues.

* Brig: Servantify `PUT /self/locale` endpoint

* Brig: Servantify `PUT /self/handle` endpoint

* Add changelog

* Delete duplicate errors use ErrorDescription where possible

* Delete unused swagger 1.x docs

* Brig: Deflake user update test

The test was using a hardcoded name for a user and searching for the user. Since
the DB is not cleaned up after/before every test run, this causes the new users
to not show up in the search results.

* Use maybe{To,From}Union instead of hand-written AsUnion instances