Servantify Cannon's internal API (#2081)
* Single notification pushes are strict in their HTTP body

The websockets library does not support streaming
(https://github.com/jaspervdj/websockets/issues/119). So, there is no
value in superficially transforming data to a stream that later won't be
streamed. However, WebSocketsData enables us to be polymorphic in
the message type, i.e. let the endpoints type decide what to use. This
would make streaming easier to enable (if it's ever implemented by the
library) and safes us from some nasty conversion code.

* Forward the pushed notification as is

This reflects the prior behavior of Cannon (that should not change).
The type RawJson represents json content as plain text.