Throw error in IO and catch it in handler to respond correctly
This error cannot be easily encoding in the servant type as it is "dynamic" from
the perspective of brig. Once we servantify galley's internal API, we can know
about this error statically. This way we can catch it and throw it as a
`RegisterError` and it will show up in brigs a `RegisterError` and it will show
up in brig's swagger.