galley/integration: Update hardcoded client ID
The client ID corresponds to the last prekeys hardcoded in the tests. It is
calculated using `Wire.API.Client.clientIdFromPrekey`. With upgrade of the
hashable package, this function now generates a different client id for a given
prekey. Not sure if this means something bad for already registered legalhold
clients.