Spar debugging (#2214)
* find user by email + refactoring

* Cleanup.

* [Experiment] try post.

This emulates scim peer behavior that makes no sense.  I was trying to
reproduce an exceedingly evasive production issue, but instead of
successfully (and erroneously) creating a broken duplicate user, this
test fails as desired with "externalId is already taken".

* Revert "[Experiment] try post."

This reverts commit 1d0f0fb44a71df468fe0cfb1948c567310f12f4d.

* Log more UserId in scim end-points (level info).

* Remove redundant log entries.

(This is already logged inside `deleteScimUser`.)

* Update user index tables in spar more exhaustively.

Do this by specializing runExternalId combinator (it was a bit vague
before).  This also has the effect of making the semantics in a few
places clearer (and more correct?).

* changelog

* Update services/spar/src/Spar/Scim/User.hs

Co-authored-by: Leif Battermann <leif.battermann@wire.com>

Co-authored-by: Leif Battermann <leif.battermann@wire.com>