Allow CORS requests at random ports for defined origin hosts (SQPIT-922)
The QA runs tests against a HTTP proxy (HA proxy). This changes the
origin in two ways:
- The ports aren't default (i.e. 80 or 443) but something more random
- The host part contains a different subdomain, e.g.
  custom.qa-fixed-sso.wire.link

To enable this behaviour for testing but keep the rules strict in
production, introduce the Helm value
nginx_conf.randomport_allowlisted_origins. This list defines to which
origins CORS requests at random ports are allowed. The default is: To
none.