Gundeck: Avoid dropping messages when redis is down (#2295)
Right now, if redis is down, the lookup of presences which happens before the storage of notifications in cassandra, leads to messages unable to be persisted in the case of redis downtime. That's not necessary. This PR moves the "lookup presence location in redis" part further down in the "pushAll" function, making gundeck's code more tolerant to redis downtimes. An integration test is provided (using a tcp proxy to redis) to cover this case.