Replace brig custom setup with file-embed-lzma
Since we are using file-embed-lzma to embed swagger definitions in the
executable (and servant-swagger itself is using it to embed the swagger
frontend), there is no point in maintaining a second way to embed static
files into the executable.

Therefore, this commit removes the mechanism for loading the swagger
description at compile time via a custom `Setup.hs`, introduced in
https://github.com/wireapp/wire-server/pull/1956, and simply embeds it
using `embedText` from `file-embed-lzma`.