Add graceful termination to coturn chart (#2456)
* charts/coturn: add support for graceful termination.

This adds a preStop hook to coturn pods which waits until the number of
active allocations goes to zero before the coturn process is signalled to
terminate. This can be used with an external service discovery mechanism to
enable graceful rolling restarts, by removing each running coturn pod from the
set of active advertised coturn instances in turn.

* changelog: add entry for coturn graceful termination feature.

* charts/coturn: update coturn container image version.

Starting with this version, our coturn image ships with the pre-stop hook
script, which is required for graceful termination.