retry Redis connection in case of network errors (#2512)
The plain `hedis` client discards the initial connection data and only retains a list of Redis cluster node IPs. When none of these IPs is valid anymore, for instance, due to Redis cluster or (Kubernetes) node updates, `hedis` immediately looses all retained connections to Redis without any option to reconnect. In this patch, we wrap the `hedis` client and retry connecting with the initially provided connection data in case of network errors. Also, the wrapper makes sure that
* if reconnecting fails, it is retried with exponential back-off, and
* only one thread reconnects while all other threads are blocked and are immediately unblocked as soon as the new connection is established.