docs: merge nix code into toplevel
For now, this introduces a second nixpkgs pin to the toplevel
nix/sources.json, and exposes the doc environment as a `docsEnv`
attribute.

The binaries in there are added to $PATH by direnv.

As Dockerfile can't copy files from outside its folder, and the only
purpose of it was to provide a binaries to build as well (which we can
just provide via Nix), it's removed.

There's now an additional `.envrc` in `docs/`, adding doc-related
tooling to $PATH whenever the `docs` folder is entered.

Building can happen via a a top-level `docs` attribute, without the need
for intermediate Docker containers.

This removes the necessity for a bunch of Makefile rules and
intermediate container images, as well as additional CI recipies - we
can just build the nix attribute from GH Actions.

The GH actions pipeline is extended to upload/update docs in our docs
bucket. Permissions for this come via OIDC and the GitHub actions
tokens.

The documentation is updated to match the new state of things.