Sanitize `delete-cache-on-linker-errors` make target (#2623)
`dist-newstyle`: The directory name depends on the cabal version. `cabal
clean` does the job (using the tool itself to manage it's cache).

`~/.cabal`: We shouldn't delete the whole folder as it may contain a
manually tweaked `config` file. Deleting `~/.cabal/store` should be
sufficient.

If `$CABAL_DIR` is set, it's taken into account.

Co-authored-by: Stefan Matting <smatting@users.noreply.github.com>