Add additional checks on incoming MLS messages (#2618)
Removed the test "user tries to remove themselves from conversation". In the
test, the group creator was creating a remove commit which was then send by Bob
to the backend. It is not working anymore, since now it is made sure by the
backend that Bob creates all his commits himself. Switching to Bob creating the
remove commit for his clients won't work either, since OpenMLS would forbid to
create a commit in which the user even removes one of his/her own clients. Thus,
the test was beyond repairable. However, there might be the option to craft a
remove commit for Bob without OpenMLS later on when crafting MLS messages is
available.

Co-authored-by: Paolo Capriotti <paolo@capriotti.io>