wireServer.ciImage: Build using official nix image
Having 'runAsRoot' in a dockerTools.buildImage requires the builder to have
"kvm" feature enabled. This is not enabled in the default nix image, so the
build fails.

Without adding users and groups, we cannot build nix in the ciImage, so in order
to ensure everything works and the image has the packages we need, we use the
derivation in the github.com/NixOS/nix repository to build a docker image with nix.