Make deletions via SCIM more stable (#2637)
Cassandra doesn't support transactions. Thus, in rare circumstances, a user
could be only partially deleted in brig (e.g. due to the pod shutting down). To
be able to clean up a partially deleted user/account, the SCIM user deletion
handler now executes the internal deletion function in brig again even if the
user is not found in brig as it's only a "tombstone". This internal deletion
function then figures out if the user ever existed and if there are any left
overs. In case, deletion is executed for the user/account again.

To gather the result of a user deletion, the brig endpoint is now synchronous
(was asynchronous before).

Co-authored-by: Matthias Fischmann <mf@zerobuzz.net>