Feature flag for exposing invite URLs to team admins [SQPIT-1368] (#2684)
This flag is generally locked and disabled. To be able to unlock/enable it, the team id must be added to the list of eligible teams. 

The invitation URLs are meant to be used in automatized onboarding processes; e.g. rendering them as QR codes.

N.B. getting hold of the invitation URLs makes it possible to impersonate as someone else. That's why there's this two-configs solution: To make sure it's only activated when absolutely needed (and not accidentally.)

Co-authored-by: Sven Tennie <sven.tennie@wire.com>
Co-authored-by: Molly Miller <molly.miller@wire.com>
Co-authored-by: Matthias Fischmann <mf@zerobuzz.net>