AWS/SNS create endpoint without custom user data (SQCORE-1267) (#2910)
We need these trace logs to understand how AWS/SNS exactly
behaves. (Unfortunately, the documentation is a bit sparse.)

The sent CustomUserData used to be the UserId. This doesn't make sense
anymore; given that a user may be attached with the same device (Android
or iOS) and several accounts to the same backend.