bullet proof smtp (fix flaky test) (#2868)
Add more logging and ensure that the logs are written. Apply a timeout to all calls such that firewalls and SMTP servers cannot delay SMTP connections. (There are some techniques to delay conversations with spammers forever / for a long time.)

To stabilize the test setup (the first attempt to merge this branch had to be reverted): Use random sockets (and by this ports) generated by a library. Keep the sockets open while the test runs. (Preventing further race-conditions.)

Co-authored-by: fisx <mf@zerobuzz.net>
Co-authored-by: Leif Battermann <leif.battermann@wire.com>