galley: Avoid duplicate work for conversation-created notificiations
For every member in the conversation, the view of the conversation differs as
they see their own member metadata in detail which other memebrs' metadata is
not so detailed. When a conversation is created, each member of the conversation
gets their own view of the convesation in a notification. The
`convRemoteMembers` and `convLocalMembers` fields are mostly repeated across all
of these views. This commit avoids computing these lists from scratch n times,
where n is the number of members in the conversation.